+++ packages/spotlight/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Added
 
-- `spotlight/Spottable` config option `selectionKeys` to support using alternate keys to fire selection-based events.
+- `spotlight/Spottable` property `selectionKeys`
 
 ## [2.1.0] - 2018-08-20